restructured template
