+++ src/lib/kombi/Template/Procedure/ImageSequence.py
