+++ src/lib/kombi/Template/__init__.py
@@ -1,0 +1,2 @@
+from .Template import Template, TemplateError, TemplateVarNotFoundError, TemplateRequiredPathNotFoundError, TemplateProcedureNotFoundError
+from . import Procedure
