+++ tests/Template/__init__.py
@@ -1,0 +1,2 @@
+from .TemplateTest import TemplateTest
+from . import Procedure
