+++ tests/__init__.py
@@ -1,8 +1,7 @@
 from .BaseTestCase import BaseTestCase
-from .TemplateTest import TemplateTest
 from .CliTest import CliTest
 from . import Crawler
-from . import TemplateProcedure
+from . import Template
 from . import Task
 from . import TaskReporter
 from . import TaskWrapper
